# services/platform-api/src/main/java/ai/nova/platform/llm/entity/LlmModelFamily.java
package ai.nova.platform.llm.entity;

public enum LlmModelFamily {
    LLAMA_3,
    LLAMA_3_1,
    LLAMA_3_2,
    QWEN,
    DEEPSEEK,
    PHI,
    GEMMA,
    MISTRAL,
    TINYLLAMA,
    CODELLAMA,
    CUSTOM
}